Le meme piege, une colonne plus loin : les montants aussi
Une fois `$1` type, l'erreur s'est deplacee : « column montant_cents is of type
integer but expression is of type text ». Dans une union, Postgres resout le type
d'un parametre sur l'ensemble de la requete, et le pilote envoie les nombres sans
type declare : les trois montants partaient en texte vers une colonne `integer`.

Meme correction, meme raison : on dit au serveur ce qu'on lui envoie. Le montage du
cas ne s'executait toujours pas, et tant qu'il ne s'execute pas il ne prouve rien.

+++ api/db-check.js
@@ -839,15 +839,15 @@ async function main() {
                 now() - ((m.id % 20) || ' hours')::interval
            from matches m where m.user_id = $1::bigint
          union all
-         select 'gain', m.id::text, 'maison:contrepartie', 'enjeu:'||m.id, $2,
+         select 'gain', m.id::text, 'maison:contrepartie', 'enjeu:'||m.id, $2::integer,
                 now() - ((m.id % 20) || ' hours')::interval
            from matches m where m.user_id = $1::bigint
          union all
-         select 'gain', m.id::text, 'enjeu:'||m.id, 'maison:commission', $3,
+         select 'gain', m.id::text, 'enjeu:'||m.id, 'maison:commission', $3::integer,
                 now() - ((m.id % 20) || ' hours')::interval
            from matches m where m.user_id = $1::bigint
          union all
-         select 'gain', m.id::text, 'enjeu:'||m.id, 'joueur:'||$1||':disponible', $4,
+         select 'gain', m.id::text, 'enjeu:'||m.id, 'joueur:'||$1||':disponible', $4::integer,
                 now() - ((m.id % 20) || ' hours')::interval
            from matches m where m.user_id = $1::bigint`,
         [u, p.grossCents - 1000, p.feeCents, p.netCents]);
